refactor: flatten presentation widgets for smaller features
Flattened widget folder structure for debts, goals, recurring, and reports features to eliminate single-file subdirectories. Updated import paths across the codebase.
