+++ lib/features/debts/presentation/widgets/debt_card.dart
@@ -5,10 +5,10 @@ import 'package:poka_ce/app/router/router.dart';
 import 'package:poka_ce/core/enums.dart';
 import 'package:poka_ce/features/debts/domain/debt_model.dart';
 import 'package:poka_ce/features/debts/presentation/controllers/debt_list_notifier.dart';
-import 'package:poka_ce/features/debts/presentation/widgets/cards/components/debt_due_date_chip.dart';
-import 'package:poka_ce/features/debts/presentation/widgets/cards/components/debt_progress_bar.dart';
-import 'package:poka_ce/features/debts/presentation/widgets/cards/components/debt_status_chip.dart';
-import 'package:poka_ce/features/debts/presentation/widgets/forms/debt_form_sheet.dart';
+import 'package:poka_ce/features/debts/presentation/widgets/debt_due_date_chip.dart';
+import 'package:poka_ce/features/debts/presentation/widgets/debt_form_sheet.dart';
+import 'package:poka_ce/features/debts/presentation/widgets/debt_progress_bar.dart';
+import 'package:poka_ce/features/debts/presentation/widgets/debt_status_chip.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/shared/widgets/poka_amount_text.dart';
 import 'package:poka_ce/shared/widgets/poka_icon.dart';
