+++ lib/features/debts/presentation/widgets/debt_list_content.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:poka_ce/features/debts/domain/debt_model.dart';
-import 'package:poka_ce/features/debts/presentation/widgets/cards/debt_card.dart';
-import 'package:poka_ce/features/debts/presentation/widgets/cards/debt_summary_card.dart';
-import 'package:poka_ce/features/debts/presentation/widgets/forms/debt_form_sheet.dart';
+import 'package:poka_ce/features/debts/presentation/widgets/debt_card.dart';
+import 'package:poka_ce/features/debts/presentation/widgets/debt_form_sheet.dart';
+import 'package:poka_ce/features/debts/presentation/widgets/debt_summary_card.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/shared/widgets/poka_empty_view.dart';
 import 'package:poka_ce/shared/widgets/poka_section_label.dart';
