+++ lib/features/goals/presentation/screens/goal_list_page.dart
@@ -4,9 +4,9 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:poka_ce/features/goals/presentation/controllers/goal_list_view_notifier.dart';
 import 'package:poka_ce/features/goals/presentation/controllers/goal_notifier.dart';
-import 'package:poka_ce/features/goals/presentation/widgets/cards/goal_card.dart';
-import 'package:poka_ce/features/goals/presentation/widgets/cards/goal_summary_card.dart';
-import 'package:poka_ce/features/goals/presentation/widgets/forms/goal_form_sheet.dart';
+import 'package:poka_ce/features/goals/presentation/widgets/goal_card.dart';
+import 'package:poka_ce/features/goals/presentation/widgets/goal_form_sheet.dart';
+import 'package:poka_ce/features/goals/presentation/widgets/goal_summary_card.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/shared/widgets/poka_empty_view.dart';
 import 'package:poka_ce/shared/widgets/poka_header.dart';
