+++ lib/features/goals/presentation/widgets/goal_card.dart
@@ -6,8 +6,8 @@ import 'package:poka_ce/app/router/router.dart';
 import 'package:poka_ce/core/enums.dart';
 import 'package:poka_ce/features/goals/presentation/controllers/goal_detail_notifier.dart';
 import 'package:poka_ce/features/goals/presentation/controllers/goal_notifier.dart';
-import 'package:poka_ce/features/goals/presentation/widgets/badges/goal_status_badge.dart';
-import 'package:poka_ce/features/goals/presentation/widgets/forms/goal_form_sheet.dart';
+import 'package:poka_ce/features/goals/presentation/widgets/goal_form_sheet.dart';
+import 'package:poka_ce/features/goals/presentation/widgets/goal_status_badge.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/shared/widgets/poka_amount_text.dart';
 import 'package:poka_ce/shared/widgets/poka_icon.dart';
