+++ lib/features/recurring/presentation/screens/recurring_list_page.dart
@@ -10,9 +10,9 @@ import 'package:flutter_animate/flutter_animate.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:poka_ce/features/recurring/domain/recurring_model.dart';
 import 'package:poka_ce/features/recurring/presentation/controllers/recurring_list_notifier.dart';
-import 'package:poka_ce/features/recurring/presentation/widgets/cards/recurring_card.dart';
-import 'package:poka_ce/features/recurring/presentation/widgets/cards/recurring_summary_card.dart';
-import 'package:poka_ce/features/recurring/presentation/widgets/forms/recurring_form_sheet.dart';
+import 'package:poka_ce/features/recurring/presentation/widgets/recurring_card.dart';
+import 'package:poka_ce/features/recurring/presentation/widgets/recurring_form_sheet.dart';
+import 'package:poka_ce/features/recurring/presentation/widgets/recurring_summary_card.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/shared/widgets/poka_empty_view.dart';
 import 'package:poka_ce/shared/widgets/poka_header.dart';
