+++ lib/features/recurring/presentation/widgets/recurring_card.dart
@@ -7,10 +7,10 @@ import 'package:poka_ce/app/router/router.dart';
 import 'package:poka_ce/core/enums.dart';
 import 'package:poka_ce/features/recurring/domain/recurring_model.dart';
 import 'package:poka_ce/features/recurring/presentation/controllers/recurring_list_notifier.dart';
-import 'package:poka_ce/features/recurring/presentation/widgets/chips/recurring_next_date_chip.dart';
-import 'package:poka_ce/features/recurring/presentation/widgets/chips/recurring_period_chip.dart';
-import 'package:poka_ce/features/recurring/presentation/widgets/chips/recurring_status_chip.dart';
-import 'package:poka_ce/features/recurring/presentation/widgets/forms/recurring_form_sheet.dart';
+import 'package:poka_ce/features/recurring/presentation/widgets/recurring_form_sheet.dart';
+import 'package:poka_ce/features/recurring/presentation/widgets/recurring_next_date_chip.dart';
+import 'package:poka_ce/features/recurring/presentation/widgets/recurring_period_chip.dart';
+import 'package:poka_ce/features/recurring/presentation/widgets/recurring_status_chip.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/shared/widgets/dialogs/poka_confirm_dialog.dart';
 import 'package:poka_ce/shared/widgets/poka_amount_text.dart';
