+++ test/feature/features/reports/presentation/widgets/allocation_row_tile_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:forui/forui.dart';
-import 'package:poka_ce/features/reports/presentation/widgets/tiles/allocation_row_tile.dart';
+import 'package:poka_ce/features/reports/presentation/widgets/allocation_row_tile.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/theme/theme.dart';
