+++ test/feature/features/reports/presentation/widgets/report_budget_utilization_test.dart
@@ -6,7 +6,8 @@ import 'package:poka_ce/core/enums.dart';
 import 'package:poka_ce/features/budgets/domain/budget_model.dart';
 import 'package:poka_ce/features/budgets/presentation/controllers/budget_progress_provider.dart';
 import 'package:poka_ce/features/dashboard/presentation/controllers/balance_visibility_provider.dart';
-import 'package:poka_ce/features/reports/presentation/widgets/tiles/budget_item_tile.dart';
+import 'package:poka_ce/features/reports/presentation/controllers/report_notifier.dart';
+import 'package:poka_ce/features/reports/presentation/widgets/report_budget_utilization.dart';
 import 'package:poka_ce/features/settings/presentation/controllers/settings_notifier.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/theme/theme.dart';
@@ -15,9 +16,9 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
   setUpAll(() => LocaleSettings.setLocaleSync(AppLocale.en));
 
-  BudgetModel budget(int amount) => BudgetModel(
-    id: 'b1',
-    name: 'Food',
+  BudgetModel budget(String id, String name, int amount) => BudgetModel(
+    id: id,
+    name: name,
     amount: amount,
     period: BudgetPeriod.monthly,
     startDate: DateTime(2026, 1, 1),
@@ -25,9 +26,15 @@ void main() {
     updatedAt: DateTime(2026, 1, 1),
   );
 
-  Widget wrap({required int spent, required int limit}) {
+  Widget wrap({
+    required List<BudgetModel> budgets,
+    int spent = 0,
+  }) {
     return ProviderScope(
       overrides: [
+        reportProvider.overrideWithValue(
+          ReportState(isLoading: false, budgets: budgets),
+        ),
         budgetProgressProvider.overrideWith((ref, arg) async => spent),
         balanceVisibilityProvider.overrideWithValue(true),
         settingsProvider.overrideWith(() => _FakeSettingsNotifier()),
@@ -35,10 +42,10 @@ void main() {
       child: TranslationProvider(
         child: MaterialApp(
           builder: (context, child) => FTheme(data: lightTheme, child: child!),
-          home: Scaffold(
-            body: Padding(
-              padding: const EdgeInsets.all(16),
-              child: BudgetItemTile(budget: budget(limit)),
+          home: const Scaffold(
+            body: SingleChildScrollView(
+              padding: EdgeInsets.all(16),
+              child: ReportBudgetUtilization(),
             ),
           ),
         ),
@@ -46,25 +53,31 @@ void main() {
     );
   }
 
-  group('BudgetItemTile', () {
-    testWidgets('shows progress percentage when under budget', (tester) async {
-      await tester.pumpWidget(wrap(spent: 500, limit: 1000));
+  group('ReportBudgetUtilization', () {
+    testWidgets('shows empty state when no budgets configured', (tester) async {
+      await tester.pumpWidget(wrap(budgets: []));
       await tester.pumpAndSettle();
 
-      expect(find.text('Food'), findsOneWidget);
+      expect(find.text('No budgets configured'), findsOneWidget);
+      expect(find.text('Add budgets to track your spending limits'), findsOneWidget);
+    });
+
+    testWidgets('renders overall utilization and individual budget tiles', (tester) async {
+      await tester.pumpWidget(
+        wrap(budgets: [budget('b1', 'Groceries', 1000)], spent: 500),
+      );
+      await tester.pumpAndSettle();
+
+      expect(find.text('Groceries'), findsOneWidget);
+      expect(find.textContaining('50.0% Spent'), findsOneWidget);
+      expect(find.text('Remaining'), findsOneWidget);
       expect(find.text('50%'), findsOneWidget);
-      expect(find.text('Over budget'), findsNothing);
     });
 
-    testWidgets('shows warning style when at 80%', (tester) async {
-      await tester.pumpWidget(wrap(spent: 800, limit: 1000));
-      await tester.pumpAndSettle();
-
-      expect(find.text('80%'), findsOneWidget);
-    });
-
-    testWidgets('shows over budget chip when spending exceeds limit', (tester) async {
-      await tester.pumpWidget(wrap(spent: 1500, limit: 1000));
+    testWidgets('marks overall progress as danger when over budget', (tester) async {
+      await tester.pumpWidget(
+        wrap(budgets: [budget('b1', 'Groceries', 1000)], spent: 1500),
+      );
       await tester.pumpAndSettle();
 
       expect(find.text('Over budget'), findsOneWidget);
