+++ test/feature/features/reports/presentation/widgets/report_cashflow_chart_test.dart
@@ -4,7 +4,7 @@ import 'package:forui/forui.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:poka_ce/features/reports/domain/services/report_analytics_service.dart';
 import 'package:poka_ce/features/reports/presentation/controllers/report_notifier.dart';
-import 'package:poka_ce/features/reports/presentation/widgets/charts/report_cashflow_chart.dart';
+import 'package:poka_ce/features/reports/presentation/widgets/report_cashflow_chart.dart';
 import 'package:poka_ce/i18n/strings.g.dart';
 import 'package:poka_ce/theme/theme.dart';
